requirements.txt
aiogram>=3.4,<4
SQLAlchemy>=2.0,<3
asyncpg>=0.28,<1
alembic>=1.13,<2
python-dotenv>=1.0,<2
aiohttp>=3.9,<4
